Add files via upload
This is a functional Winsock IPv4, IPv6 client and server:
This was a winsock api I built on blocking sockets as a learning purpose, and to assist others in venturing in to their goals of making a online system of some form.  This is a very generic blocking socket System and will likely require additional modification to suit the needs and purposes of your uses.